Add a really simple way to sync directories.
Basically, omni.conf has both remote and local dirs. It also has the idea
of remote computers with e.g. addresses. So it is easy to use this info to do an
rsync between the two computers. It will not work in the case when there are lots
of possibly unneeded files on the remote computer though. It also needs to be
integrated with the remote-node-sync command somehow too. As a quick hack it's
pretty nice though.