Close #24: Add pylint_tests to tests.
pylint is pretty strict: don't try to clear up all warnings yet.
Test checks that the pylint score is over a certain value (6 currently for src).
And that there are no errors (some had to be disabled with # pylint: disable=...).
Has already helped to catch some issues in the code, should be useful in future.