Logic for distributing tasks more or less there.
TaskMaster now doles out tasks when there are pending tasks with met deps, and None where then are not, raising a StopIteration when all tasks are either completed or being worked on.
MpiMaster sends tasks to slaves based on this, clearing a backlog when tasks are blocked on a given task completing.