Work on converter/deleter.
Turn FF2NC_Converter into a subclass of Analysis so that it can be called as such.
Start thinking about how deleter class should work. Complicated by fact that it
will remove files from data_dir, making it harder to make it play nicely with current
setup. Add in a min/max runid config option for analysis, which should be useful for
deleter.