Add to scripts for running on ARCHER without stepping on rose/cylc's toes.
rose and cylc are Python2 only. This causes problems for me when running with Py3 omnium.
The solution is to add an extra layer of indirection - instead of setting up Py3 (with iris)
then calling omnium directly from cylc or rose, call a bash script which sets up Py3 and then
launches omnium. This effectively shields the calling PBS script from the `module load`s performed.
The two scripts added are the scripts that get called. In the future, these should be made more
generic. Either omnium/bin should be added to PATH, or these should be copied to PATH (e.g.
<cylc_suite>/bin).