Wide ranging changes to how analysis_pkgs are imported.
The idea is that instead of having one version of the analysis on each
computer, there is one version cloned into each suite. All the info
to clone it and checkout a specific version is in app/omnium/rose-app.conf
and updating is done by `omnium setup-analysis`, which will clone the
specific branch/commit/tag. It is cloned into `.omnium/analysis_pkgs/`,
which is added to sys.path when finding analysers. Every other command
checks that commit matches the current commit, and gives you a message to
run the setup command if it doesn't. This might happen if e.g. you
are checking out master, and the master it changed (it does a `git fetch`
each time it runs).